Add XMLType Package
- .ecore files which have been created from a .xsd file use the
  schema http://www.eclipse.org/emf/2003/XMLType, when they
  refer to primitive datatypes.

- emf4cpp already supports "external" packages. It uses the
  package name and the type name from the external package, when
  a datatype is used. It also include a forward definition header
  file as needed.

- However, the initialization of the metaclasses in the generated
  code require the existence of the metaclasses of the used
  packages. This can be reached best, when the used packages has
  also been created by emf4cpp.

- Therefor this patch adds the XMLType.ecore definition file of
  the schema http://www.eclipse.org/emf/2003/XMLType. It has been
  bootstrapped, the created code has been added, and it has
  integrated in the build process.

- The installation directory is the same as for the ecore library
  itself, so it can be found automatically.

- Current drawback: emf4cpp uses the package name as the C++
  namespace. For XMLType.ecore this is simply "type". This name
  may conflict with client code because it is too
  general. XMLType.ecore also specifies a namespace prefix
  "ecore.xml.type" which could be used to give better (nested)
  namespaces. This should be handled in a different issue.

- Added (defunctional) string_trait for serialization of
  EList<EString>.

- Changed License notice, introduced emf4cpp@inchron.com as an
  email alias.

- Regenerated library implementation and test projects.

+++ emf4cpp.tests/kdm/kdm/core_forward.hpp
@@ -2,7 +2,7 @@
 /*
  * kdm/core_forward.hpp
  * Copyright (C) Cátedra SAES-UMU 2010 <andres.senac@um.es>
- * Copyright (C) INCHRON GmbH 2016 <soeren.henning@inchron.com>
+ * Copyright (C) INCHRON GmbH 2016-2018 <emf4cpp@inchron.com>
  *
  * EMF4CPP is free software: you can redistribute it and/or modify it
  * under the terms of the GNU Lesser General Public License as published
@@ -45,7 +45,7 @@ namespace kdm
         /*PROTECTED REGION ID(kdm_core_Integer) START*/
 // Please, enable the protected region if you add manually written code.
 // To do this, add the keyword ENABLED before START.
-        typedef int Integer;
+        typedef int64_t Integer;
         /*PROTECTED REGION END*/
         /*PROTECTED REGION ID(kdm_core_Boolean) ENABLED START*/
         typedef bool Boolean;
